generador_trafico/requirements.txt
pandas
numpy
requests
kafka-python
psycopg2-binary